11/18 수정
