+++ src/main/java/com/spring/model/RecommendationDto.java
@@ -7,6 +7,6 @@ import java.time.LocalDate;
 import java.time.LocalDateTime;
 
 
-public record RecommendationDto(int id, String userId,String recStyle, String recActivity, String temp_high, String temp_low, LocalDateTime createDate) {
+public record RecommendationDto(int id, String userId,String recStyle, String recActivity, String temp_high, String temp_low, LocalDateTime createDate, String imageUrl) {
 
 }
