+++ src/main/java/com/spring/repository/social/RecommendationRepository.java
@@ -10,6 +10,5 @@ import java.util.List;
 
 public interface RecommendationRepository extends JpaRepository<Recommendation, Long> {
   List<Recommendation> findByUserId_UserId(String userId);
-
   void deleteByUserId_UserId(String userId);
 }
